App.vue: add vue router split components
