+++ src/main.ts
@@ -1,7 +1,10 @@
 import { createApp } from 'vue';
 import '@/assets/css/main.css';
 import App from './App.vue';
+import router from './router';
 
 const app = createApp(App);
+
+app.use(router);
 
 app.mount('#app');
