+++ src/views/layouts/HeaderLayout.vue
@@ -1,0 +1,183 @@
+<script setup lang="ts">
+    import { onMounted } from 'vue';
+
+    onMounted(() => {
+        if (localStorage.getItem('theme') === 'dark' || (!('theme' in localStorage) && window.matchMedia('(prefers-color-scheme: dark)').matches)) {
+            document.documentElement.classList.add('dark');
+        } else {
+            document.documentElement.classList.remove('dark');
+        }
+    });
+
+    const toggleTheme = (): void => {
+        if (document.documentElement.classList.contains('dark')) {
+            document.documentElement.classList.remove('dark');
+            localStorage.setItem('theme', 'light');
+        } else {
+            document.documentElement.classList.add('dark');
+            localStorage.setItem('theme', 'dark');
+        }
+    };
+</script>
+
+<template>
+    <header class="contents lg:pointer-events-none lg:fixed lg:inset-0 lg:z-40 lg:flex">
+        <div class="contents lg:pointer-events-auto lg:block lg:w-72 lg:overflow-y-auto lg:border-r lg:border-gray-light lg:px-6 lg:pb-8 lg:pt-3 lg:dark:border-white/10 xl:w-80">
+            <div class="hidden lg:flex">
+                <a aria-label="Home" href="/">
+                    <span class="text-primary dark:text-white text-2xl">FlowForge</span>
+                </a>
+            </div>
+            <div class="fixed inset-x-0 top-0 z-50 flex h-14 items-center justify-between gap-12 px-4 border-b border-gray-light dark:border-gray-dark transition sm:px-6 lg:left-72 lg:z-30 lg:px-8 xl:left-80 backdrop-blur-sm dark:backdrop-blur bg-white/[var(--bg-opacity-light)] dark:bg-zinc-800/[var(--bg-opacity-dark)]">
+                <div class="absolute inset-x-0 top-full h-px transition bg-zinc-900/7.5 dark:bg-white/7.5"></div>
+                <div class="hidden lg:block lg:max-w-md lg:flex-auto">
+                    <div class="relative">
+                        <svg viewBox="0 0 20 20" fill="none" aria-hidden="true" class="absolute left-2 top-1/2 h-5 w-5 stroke-current transform -translate-y-1/2">
+                            <path stroke-linecap="round" stroke-linejoin="round" d="M12.01 12a4.25 4.25 0 1 0-6.02-6 4.25 4.25 0 0 0 6.02 6Zm0 0 3.24 3.25"></path>
+                        </svg>
+                        <input type="search" class="h-8 w-full rounded-xl bg-white pl-10 pr-3 text-sm text-zinc-500 ring-1 ring-zinc-900/10 transition focus:outline-none focus:ring-1 focus:ring-gray-light dark:bg-white/5 dark:text-zinc-400 dark:ring-inset dark:ring-white/10 lg:flex" placeholder="Find something...">
+                    </div>
+                </div>
+                <div class="flex items-center gap-5 lg:hidden">
+                    <button type="button" class="flex h-6 w-6 items-center justify-center rounded-md transition hover:bg-zinc-900/5 dark:hover:bg-white/5" aria-label="Toggle navigation">
+                        <svg viewBox="0 0 10 9" fill="none" stroke-linecap="round" aria-hidden="true" class="w-2.5 stroke-zinc-900 dark:stroke-white">
+                            <path d="M.5 1h9M.5 8h9M.5 4.5h9"></path>
+                        </svg>
+                    </button>
+                    <a aria-label="Home" href="/">
+                        <svg viewBox="0 0 99 24" aria-hidden="true" class="h-6">
+                            <path class="fill-primary" d="M16 8a5 5 0 0 0-5-5H5a5 5 0 0 0-5 5v13.927a1 1 0 0 0 1.623.782l3.684-2.93a4 4 0 0 1 2.49-.87H11a5 5 0 0 0 5-5V8Z"></path>
+                            <path class="fill-zinc-900 dark:fill-white" d="M26.538 18h2.654v-3.999h2.576c2.672 0 4.456-1.723 4.456-4.333V9.65c0-2.61-1.784-4.333-4.456-4.333h-5.23V18Zm4.58-10.582c1.52 0 2.416.8 2.416 2.241v.018c0 1.441-.896 2.25-2.417 2.25h-1.925V7.418h1.925ZM38.051 18h2.566v-5.414c0-1.371.923-2.206 2.382-2.206.396 0 .791.061 1.178.15V8.287a3.843 3.843 0 0 0-.958-.123c-1.257 0-2.136.615-2.443 1.661h-.159V8.323h-2.566V18Zm11.55.202c2.979 0 4.772-1.88 4.772-5.036v-.018c0-3.128-1.82-5.036-4.773-5.036-2.953 0-4.772 1.916-4.772 5.036v.018c0 3.146 1.793 5.036 4.772 5.036Zm0-2.013c-1.372 0-2.145-1.116-2.145-3.023v-.018c0-1.89.782-3.023 2.144-3.023 1.354 0 2.145 1.134 2.145 3.023v.018c0 1.907-.782 3.023-2.145 3.023Zm10.52 1.846c.492 0 .967-.053 1.283-.114v-1.907a6.057 6.057 0 0 1-.755.044c-.87 0-1.24-.387-1.24-1.257v-4.544h1.995V8.323H59.41V6.012h-2.592v2.311h-1.495v1.934h1.495v5.133c0 1.88.949 2.645 3.304 2.645Zm7.287.167c2.98 0 4.772-1.88 4.772-5.036v-.018c0-3.128-1.82-5.036-4.772-5.036-2.954 0-4.773 1.916-4.773 5.036v.018c0 3.146 1.793 5.036 4.773 5.036Zm0-2.013c-1.372 0-2.145-1.116-2.145-3.023v-.018c0-1.89.782-3.023 2.145-3.023 1.353 0 2.144 1.134 2.144 3.023v.018c0 1.907-.782 3.023-2.144 3.023Zm10.767 2.013c2.522 0 4.034-1.353 4.297-3.463l.01-.053h-2.374l-.017.036c-.229.966-.853 1.467-1.908 1.467-1.37 0-2.135-1.08-2.135-3.04v-.018c0-1.934.755-3.006 2.135-3.006 1.099 0 1.74.615 1.908 1.556l.008.017h2.391v-.026c-.228-2.162-1.749-3.56-4.315-3.56-3.033 0-4.738 1.837-4.738 5.019v.017c0 3.217 1.714 5.054 4.738 5.054Zm10.257 0c2.98 0 4.772-1.88 4.772-5.036v-.018c0-3.128-1.82-5.036-4.772-5.036-2.953 0-4.773 1.916-4.773 5.036v.018c0 3.146 1.793 5.036 4.773 5.036Zm0-2.013c-1.371 0-2.145-1.116-2.145-3.023v-.018c0-1.89.782-3.023 2.145-3.023 1.353 0 2.144 1.134 2.144 3.023v.018c0 1.907-.782 3.023-2.144 3.023ZM95.025 18h2.566V4.623h-2.566V18Z"></path>
+                        </svg>
+                    </a>
+                </div>
+                <div class="flex items-center gap-5">
+                    <nav class="hidden md:block">
+                        <ul role="list" class="flex items-center gap-8">
+                            <li><a class="text-sm leading-5 text-zinc-600 transition hover:text-zinc-900 dark:text-zinc-400 dark:hover:text-white" href="/">API</a></li>
+                            <li><a class="text-sm leading-5 text-zinc-600 transition hover:text-zinc-900 dark:text-zinc-400 dark:hover:text-white" href="#">Documentation</a></li>
+                            <li><a class="text-sm leading-5 text-zinc-600 transition hover:text-zinc-900 dark:text-zinc-400 dark:hover:text-white" href="#">Support</a></li>
+                        </ul>
+                    </nav>
+                    <div class="hidden md:block md:h-5 md:w-px md:bg-gray-light md:dark:bg-gray-dark"></div>
+                    <div class="flex gap-4">
+                        <div class="contents lg:hidden">
+                            <button type="button" class="flex h-6 w-6 items-center justify-center rounded-md transition hover:bg-zinc-900/5 ui-not-focus-visible:outline-none dark:hover:bg-white/5 lg:hidden" aria-label="Find something...">
+                                <svg viewBox="0 0 20 20" fill="none" aria-hidden="true" class="h-5 w-5 stroke-zinc-900 dark:stroke-white">
+                                    <path stroke-linecap="round" stroke-linejoin="round" d="M12.01 12a4.25 4.25 0 1 0-6.02-6 4.25 4.25 0 0 0 6.02 6Zm0 0 3.24 3.25"></path>
+                                </svg>
+                            </button>
+                        </div>
+                        <button type="button" class="flex h-6 w-6 items-center justify-center rounded-md transition hover:bg-zinc-900/5 dark:hover:bg-white/5" aria-label="Toggle theme" @click="toggleTheme">
+                            <svg viewBox="0 0 20 20" fill="none" aria-hidden="true" class="h-5 w-5 stroke-zinc-900 dark:hidden">
+                                <path d="M12.5 10a2.5 2.5 0 1 1-5 0 2.5 2.5 0 0 1 5 0Z"></path>
+                                <path stroke-linecap="round" d="M10 5.5v-1M13.182 6.818l.707-.707M14.5 10h1M13.182 13.182l.707.707M10 15.5v-1M6.11 13.889l.708-.707M4.5 10h1M6.11 6.111l.708.707"></path>
+                            </svg>
+                            <svg viewBox="0 0 20 20" fill="none" aria-hidden="true" class="hidden h-5 w-5 stroke-white dark:block">
+                                <path d="M15.224 11.724a5.5 5.5 0 0 1-6.949-6.949 5.5 5.5 0 1 0 6.949 6.949Z"></path>
+                            </svg>
+                        </button>
+                    </div>
+                    <div class="hidden min-[416px]:contents">
+                        <a class="inline-flex gap-0.5 justify-center overflow-hidden text-sm font-medium transition rounded-xl ring-1 ring-primary pb-1 pt-0.5 px-3 text-primary hover:bg-primary hover:text-white dark:text-white dark:ring-1 dark:ring-white dark:hover:bg-gray-light dark:hover:text-black dark:hover:ring-gray-light" href="#">Sign in</a>
+                    </div>
+                </div>
+            </div>
+            <nav class="hidden lg:mt-16 lg:block">
+                <ul role="list">
+                    <li class="relative mt-6 md:mt-0">
+                        <h2 class="text-xs font-semibold text-gray-dark dark:text-gray-light">Navigation</h2>
+                        <div class="relative mt-3 pl-2">
+                            <div class="absolute inset-x-0 top-0 bg-zinc-800/2.5 will-change-transform dark:bg-white/2.5" style="border-radius:8px;height:32px;top:0;opacity:0"></div>
+                            <div class="absolute inset-y-0 left-2 w-px bg-zinc-900/10 dark:bg-white/5"></div>
+                            <div class="absolute left-2 h-6 w-px bg-black dark:bg-white" style="top:4px;opacity:1"></div>
+                            <ul role="list" class="border-l border-transparent">
+                                <li class="relative">
+                                    <a aria-current="page" class="flex justify-between gap-2 py-1 pr-3 text-sm transition pl-4 text-zinc-900 dark:text-white" href="/">
+                                        <span class="truncate">Dashboard</span>
+                                    </a>
+                                </li>
+                                <li class="relative">
+                                    <a aria-current="page" class="flex justify-between gap-2 py-1 pr-3 text-sm transition pl-4 text-zinc-900 dark:text-white" href="/teams">
+                                        <span class="truncate">Team</span>
+                                    </a>
+                                    <ul role="list" style="opacity:1">
+                                        <li>
+                                            <a class="flex justify-between gap-2 py-1 pr-3 text-sm transition pl-7 text-zinc-600 hover:text-zinc-900 dark:text-zinc-400 dark:hover:text-white" href="/#all_members">
+                                                <span class="truncate">All Members</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a class="flex justify-between gap-2 py-1 pr-3 text-sm transition pl-7 text-zinc-600 hover:text-zinc-900 dark:text-zinc-400 dark:hover:text-white" href="/#member_groups">
+                                                <span class="truncate">Member Groups</span>
+                                            </a>
+                                        </li>
+                                    </ul>
+                                </li>
+                                <li class="relative">
+                                    <a aria-current="page" class="flex justify-between gap-2 py-1 pr-3 text-sm transition pl-4 text-zinc-900 dark:text-white" href="/projects">
+                                        <span class="truncate">Projects</span>
+                                    </a>
+                                    <ul role="list" style="opacity:1">
+                                        <li>
+                                            <a class="flex justify-between gap-2 py-1 pr-3 text-sm transition pl-7 text-zinc-600 hover:text-zinc-900 dark:text-zinc-400 dark:hover:text-white" href="/#all_projects">
+                                                <span class="truncate">All Projects</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a class="flex justify-between gap-2 py-1 pr-3 text-sm transition pl-7 text-zinc-600 hover:text-zinc-900 dark:text-zinc-400 dark:hover:text-white" href="/#project_categories">
+                                                <span class="truncate">Project Categories</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a class="flex justify-between gap-2 py-1 pr-3 text-sm transition pl-7 text-zinc-600 hover:text-zinc-900 dark:text-zinc-400 dark:hover:text-white" href="/#project_categories">
+                                                <span class="truncate">Assigned to Me</span>
+                                            </a>
+                                        </li>
+                                    </ul>
+                                </li>
+                                <li class="relative">
+                                    <a aria-current="page" class="flex justify-between gap-2 py-1 pr-3 text-sm transition pl-4 text-zinc-900 dark:text-white" href="/issues">
+                                        <span class="truncate">Issues</span>
+                                    </a>
+                                    <ul role="list" style="opacity:1">
+                                        <li>
+                                            <a class="flex justify-between gap-2 py-1 pr-3 text-sm transition pl-7 text-zinc-600 hover:text-zinc-900 dark:text-zinc-400 dark:hover:text-white" href="/#all_issues">
+                                                <span class="truncate">All Issues</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a class="flex justify-between gap-2 py-1 pr-3 text-sm transition pl-7 text-zinc-600 hover:text-zinc-900 dark:text-zinc-400 dark:hover:text-white" href="/#all_issues">
+                                                <span class="truncate">Assigned to Me</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a class="flex justify-between gap-2 py-1 pr-3 text-sm transition pl-7 text-zinc-600 hover:text-zinc-900 dark:text-zinc-400 dark:hover:text-white" href="/#open_issues">
+                                                <span class="truncate">Open Issues</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a class="flex justify-between gap-2 py-1 pr-3 text-sm transition pl-7 text-zinc-600 hover:text-zinc-900 dark:text-zinc-400 dark:hover:text-white" href="/#in_progress_issues">
+                                                <span class="truncate">In Progress Issues</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a class="flex justify-between gap-2 py-1 pr-3 text-sm transition pl-7 text-zinc-600 hover:text-zinc-900 dark:text-zinc-400 dark:hover:text-white" href="/#closed_issues">
+                                                <span class="truncate">Closed Issues</span>
+                                            </a>
+                                        </li>
+                                    </ul>
+                                </li>
+                            </ul>
+                        </div>
+                    </li>
+                    <li class="sticky bottom-0 z-10 mt-6 min-[416px]:hidden">
+                        <a class="inline-flex gap-0.5 justify-center overflow-hidden text-sm font-medium transition rounded-full bg-zinc-900 py-1 px-3 text-white hover:bg-zinc-700 dark:bg-primary dark:text-white dark:hover:bg-primary w-full" href="#">
+                            Sign Out
+                        </a>
+                    </li>
+                </ul>
+            </nav>
+        </div>
+    </header>
+</template>
