+++ src/views/pages/DashboardPage.vue
@@ -1,0 +1,164 @@
+<template>
+    <main class="flex-auto">
+        <article class="flex h-full flex-col pb-10 pt-12">
+            <div class="flex-auto prose dark:prose-invert [html_:where(&amp;&gt;*)]:mx-auto [html_:where(&amp;&gt;*)]:max-w-2xl [html_:where(&amp;&gt;*)]:lg:mx-[calc(50%-min(50%,theme(maxWidth.lg)))] [html_:where(&amp;&gt;*)]:lg:max-w-3xl">
+                <div class="xl:max-w-none">
+                    <h2 id="resources" class="scroll-mt-24">
+                        <a class="group text-inherit no-underline hover:text-inherit dark:text-white" href="#projects">Projects</a>
+                    </h2>
+                    <div class="not-prose mt-4 grid grid-cols-1 gap-8 border-t border-gray-light pt-10 dark:border-gray-dark sm:grid-cols-2 xl:grid-cols-4">
+                        <div class="group relative flex rounded-2xl border border-gray-light dark:border-gray-dark transition-shadow hover:shadow-md hover:shadow-zinc-900/5 dark:bg-white/2.5 dark:hover:shadow-black/5">
+                            <div class="relative rounded-2xl px-4 pb-4 pt-16">
+                                <div class="flex h-9 w-9 items-center justify-center rounded-full bg-white ring-1 ring-gray-light dark:ring-gray-dark backdrop-blur-[2px] transition duration-300 group-hover:bg-white/50 group-hover:ring-gray-dark dark:bg-zinc-800 dark:text-zinc-400 dark:ring-white/15 dark:group-hover:bg-zinc-800 dark:group-hover:ring-gray-light">
+                                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
+                                        <path stroke-linecap="round" stroke-linejoin="round" d="M3.75 9.776c.112-.017.227-.026.344-.026h15.812c.117 0 .232.009.344.026m-16.5 0a2.25 2.25 0 00-1.883 2.542l.857 6a2.25 2.25 0 002.227 1.932H19.05a2.25 2.25 0 002.227-1.932l.857-6a2.25 2.25 0 00-1.883-2.542m-16.5 0V6A2.25 2.25 0 016 3.75h3.879a1.5 1.5 0 011.06.44l2.122 2.12a1.5 1.5 0 001.06.44H18A2.25 2.25 0 0120.25 9v.776" />
+                                    </svg>
+                                </div>
+                                <h3 class="mt-4 text-sm font-semibold leading-7 text-zinc-900 dark:text-white">
+                                    <a href="/project-1"><span class="absolute inset-0 rounded-2xl"></span>Project 1</a>
+                                </h3>
+                                <p class="mt-1 text-sm text-zinc-600 dark:text-zinc-400">Learn about the contact model and how to create, retrieve, update, delete, and list contacts.</p>
+                            </div>
+                        </div>
+                        <div class="group relative flex rounded-2xl border border-gray-light dark:border-gray-dark transition-shadow hover:shadow-md hover:shadow-zinc-900/5 dark:bg-white/2.5 dark:hover:shadow-black/5">
+                            <div class="relative rounded-2xl px-4 pb-4 pt-16">
+                                <div class="flex h-9 w-9 items-center justify-center rounded-full bg-white ring-1 ring-gray-light dark:ring-gray-dark backdrop-blur-[2px] transition duration-300 group-hover:bg-white/50 group-hover:ring-gray-dark dark:bg-zinc-800 dark:text-zinc-400 dark:ring-white/15 dark:group-hover:bg-zinc-800 dark:group-hover:ring-gray-light">
+                                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
+                                        <path stroke-linecap="round" stroke-linejoin="round" d="M3.75 9.776c.112-.017.227-.026.344-.026h15.812c.117 0 .232.009.344.026m-16.5 0a2.25 2.25 0 00-1.883 2.542l.857 6a2.25 2.25 0 002.227 1.932H19.05a2.25 2.25 0 002.227-1.932l.857-6a2.25 2.25 0 00-1.883-2.542m-16.5 0V6A2.25 2.25 0 016 3.75h3.879a1.5 1.5 0 011.06.44l2.122 2.12a1.5 1.5 0 001.06.44H18A2.25 2.25 0 0120.25 9v.776" />
+                                    </svg>
+                                </div>
+                                <h3 class="mt-4 text-sm font-semibold leading-7 text-zinc-900 dark:text-white">
+                                    <a href="/project-2"><span class="absolute inset-0 rounded-2xl"></span>Project 2</a>
+                                </h3>
+                                <p class="mt-1 text-sm text-zinc-600 dark:text-zinc-400">Learn about the contact model and how to create, retrieve, update, delete, and list contacts.</p>
+                            </div>
+                        </div>
+                        <div class="group relative flex rounded-2xl border border-gray-light dark:border-gray-dark transition-shadow hover:shadow-md hover:shadow-zinc-900/5 dark:bg-white/2.5 dark:hover:shadow-black/5">
+                            <div class="relative rounded-2xl px-4 pb-4 pt-16">
+                                <div class="flex h-9 w-9 items-center justify-center rounded-full bg-white ring-1 ring-gray-light dark:ring-gray-dark backdrop-blur-[2px] transition duration-300 group-hover:bg-white/50 group-hover:ring-gray-dark dark:bg-zinc-800 dark:text-zinc-400 dark:ring-white/15 dark:group-hover:bg-zinc-800 dark:group-hover:ring-gray-light">
+                                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
+                                        <path stroke-linecap="round" stroke-linejoin="round" d="M3.75 9.776c.112-.017.227-.026.344-.026h15.812c.117 0 .232.009.344.026m-16.5 0a2.25 2.25 0 00-1.883 2.542l.857 6a2.25 2.25 0 002.227 1.932H19.05a2.25 2.25 0 002.227-1.932l.857-6a2.25 2.25 0 00-1.883-2.542m-16.5 0V6A2.25 2.25 0 016 3.75h3.879a1.5 1.5 0 011.06.44l2.122 2.12a1.5 1.5 0 001.06.44H18A2.25 2.25 0 0120.25 9v.776" />
+                                    </svg>
+                                </div>
+                                <h3 class="mt-4 text-sm font-semibold leading-7 text-zinc-900 dark:text-white">
+                                    <a href="/project-3"><span class="absolute inset-0 rounded-2xl"></span>Project 3</a>
+                                </h3>
+                                <p class="mt-1 text-sm text-zinc-600 dark:text-zinc-400">Learn about the contact model and how to create, retrieve, update, delete, and list contacts.</p>
+                            </div>
+                        </div>
+                        <div class="group relative flex rounded-2xl border border-gray-light dark:border-gray-dark transition-shadow hover:shadow-md hover:shadow-zinc-900/5 dark:bg-white/2.5 dark:hover:shadow-black/5">
+                            <div class="relative rounded-2xl px-4 pb-4 pt-16">
+                                <div class="flex h-9 w-9 items-center justify-center rounded-full bg-white ring-1 ring-gray-light dark:ring-gray-dark backdrop-blur-[2px] transition duration-300 group-hover:bg-white/50 group-hover:ring-gray-dark dark:bg-zinc-800 dark:text-zinc-400 dark:ring-white/15 dark:group-hover:bg-zinc-800 dark:group-hover:ring-gray-light">
+                                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
+                                        <path stroke-linecap="round" stroke-linejoin="round" d="M3.75 9.776c.112-.017.227-.026.344-.026h15.812c.117 0 .232.009.344.026m-16.5 0a2.25 2.25 0 00-1.883 2.542l.857 6a2.25 2.25 0 002.227 1.932H19.05a2.25 2.25 0 002.227-1.932l.857-6a2.25 2.25 0 00-1.883-2.542m-16.5 0V6A2.25 2.25 0 016 3.75h3.879a1.5 1.5 0 011.06.44l2.122 2.12a1.5 1.5 0 001.06.44H18A2.25 2.25 0 0120.25 9v.776" />
+                                    </svg>
+                                </div>
+                                <h3 class="mt-4 text-sm font-semibold leading-7 text-zinc-900 dark:text-white">
+                                    <a href="/project-4"><span class="absolute inset-0 rounded-2xl"></span>Project 4</a>
+                                </h3>
+                                <p class="mt-1 text-sm text-zinc-600 dark:text-zinc-400">Learn about the contact model and how to create, retrieve, update, delete, and list contacts.</p>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+                <div class="my-16 xl:max-w-none">
+                    <h2 id="guides" class="scroll-mt-24">
+                        <a class="group text-inherit no-underline hover:text-inherit dark:text-gray-light" href="#guides">Team</a>
+                    </h2>
+                    <div class="not-prose mt-4 border-t border-gray-light pt-10 dark:border-gray-dark">
+                        <ul role="list" class="divide-y divide-gray-light dark:divide-gray-dark">
+                            <li class="flex justify-between gap-x-6 py-5">
+                                <div class="flex min-w-0 gap-x-4">
+                                    <img class="h-12 w-12 flex-none rounded-full bg-gray-50" src="https://images.unsplash.com/photo-1494790108377-be9c29b29330?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80" alt="">
+                                    <div class="min-w-0 flex-auto">
+                                        <p class="text-sm font-semibold leading-6 text-gray-900 dark:text-gray-light">Leslie Alexander</p>
+                                        <p class="mt-1 truncate text-xs leading-5 text-zinc-900 dark:text-zinc-400">leslie.alexander@example.com</p>
+                                    </div>
+                                </div>
+                                <div class="hidden shrink-0 sm:flex sm:flex-col sm:items-end">
+                                    <p class="text-sm leading-6 text-gray-900 dark:text-gray-light">Co-Founder / CEO</p>
+                                    <p class="mt-1 text-xs leading-5 text-zinc-900 dark:text-zinc-400">Last seen <time datetime="2023-01-23T13:23Z">3h ago</time></p>
+                                </div>
+                            </li>
+                            <li class="flex justify-between gap-x-6 py-5">
+                                <div class="flex min-w-0 gap-x-4">
+                                    <img class="h-12 w-12 flex-none rounded-full bg-gray-50" src="https://images.unsplash.com/photo-1519244703995-f4e0f30006d5?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80" alt="">
+                                    <div class="min-w-0 flex-auto">
+                                        <p class="text-sm font-semibold leading-6 text-gray-900 dark:text-gray-light">Michael Foster</p>
+                                        <p class="mt-1 truncate text-xs leading-5 text-zinc-900 dark:text-zinc-400">michael.foster@example.com</p>
+                                    </div>
+                                </div>
+                                <div class="hidden shrink-0 sm:flex sm:flex-col sm:items-end">
+                                    <p class="text-sm leading-6 text-gray-900 dark:text-gray-light">Co-Founder / CTO</p>
+                                    <p class="mt-1 text-xs leading-5 text-zinc-900 dark:text-zinc-400">Last seen <time datetime="2023-01-23T13:23Z">3h ago</time></p>
+                                </div>
+                            </li>
+                            <li class="flex justify-between gap-x-6 py-5">
+                                <div class="flex min-w-0 gap-x-4">
+                                    <img class="h-12 w-12 flex-none rounded-full bg-gray-50" src="https://images.unsplash.com/photo-1506794778202-cad84cf45f1d?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80" alt="">
+                                    <div class="min-w-0 flex-auto">
+                                        <p class="text-sm font-semibold leading-6 text-gray-900 dark:text-gray-light">Dries Vincent</p>
+                                        <p class="mt-1 truncate text-xs leading-5 text-zinc-900 dark:text-zinc-400">dries.vincent@example.com</p>
+                                    </div>
+                                </div>
+                                <div class="hidden shrink-0 sm:flex sm:flex-col sm:items-end">
+                                    <p class="text-sm leading-6 text-gray-900 dark:text-gray-light">Business Relations</p>
+                                    <div class="mt-1 flex items-center gap-x-1.5">
+                                        <div class="flex-none rounded-full bg-emerald-500/20 p-1">
+                                            <div class="h-1.5 w-1.5 rounded-full bg-emerald-500"></div>
+                                        </div>
+                                        <p class="text-xs leading-5 text-zinc-900 dark:text-zinc-400">Online</p>
+                                    </div>
+                                </div>
+                            </li>
+                            <li class="flex justify-between gap-x-6 py-5">
+                                <div class="flex min-w-0 gap-x-4">
+                                    <img class="h-12 w-12 flex-none rounded-full bg-gray-50" src="https://images.unsplash.com/photo-1517841905240-472988babdf9?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80" alt="">
+                                    <div class="min-w-0 flex-auto">
+                                        <p class="text-sm font-semibold leading-6 text-gray-900 dark:text-gray-light">Lindsay Walton</p>
+                                        <p class="mt-1 truncate text-xs leading-5 text-zinc-900 dark:text-zinc-400">lindsay.walton@example.com</p>
+                                    </div>
+                                </div>
+                                <div class="hidden shrink-0 sm:flex sm:flex-col sm:items-end">
+                                    <p class="text-sm leading-6 text-gray-900 dark:text-gray-light">Front-end Developer</p>
+                                    <p class="mt-1 text-xs leading-5 text-zinc-900 dark:text-zinc-400">Last seen <time datetime="2023-01-23T13:23Z">3h ago</time></p>
+                                </div>
+                            </li>
+                            <li class="flex justify-between gap-x-6 py-5">
+                                <div class="flex min-w-0 gap-x-4">
+                                    <img class="h-12 w-12 flex-none rounded-full bg-gray-50" src="https://images.unsplash.com/photo-1438761681033-6461ffad8d80?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80" alt="">
+                                    <div class="min-w-0 flex-auto">
+                                        <p class="text-sm font-semibold leading-6 text-gray-900 dark:text-gray-light">Courtney Henry</p>
+                                        <p class="mt-1 truncate text-xs leading-5 text-zinc-900 dark:text-zinc-400">courtney.henry@example.com</p>
+                                    </div>
+                                </div>
+                                <div class="hidden shrink-0 sm:flex sm:flex-col sm:items-end">
+                                    <p class="text-sm leading-6 text-gray-900 dark:text-gray-light">Designer</p>
+                                    <p class="mt-1 text-xs leading-5 text-zinc-900 dark:text-zinc-400">Last seen <time datetime="2023-01-23T13:23Z">3h ago</time></p>
+                                </div>
+                            </li>
+                            <li class="flex justify-between gap-x-6 py-5">
+                                <div class="flex min-w-0 gap-x-4">
+                                    <img class="h-12 w-12 flex-none rounded-full bg-gray-50" src="https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80" alt="">
+                                    <div class="min-w-0 flex-auto">
+                                        <p class="text-sm font-semibold leading-6 text-gray-900 dark:text-gray-light">Tom Cook</p>
+                                        <p class="mt-1 truncate text-xs leading-5 text-zinc-900 dark:text-zinc-400">tom.cook@example.com</p>
+                                    </div>
+                                </div>
+                                <div class="hidden shrink-0 sm:flex sm:flex-col sm:items-end">
+                                    <p class="text-sm leading-6 text-gray-900 dark:text-gray-light">Director of Product</p>
+                                    <div class="mt-1 flex items-center gap-x-1.5">
+                                        <div class="flex-none rounded-full bg-emerald-500/20 p-1">
+                                            <div class="h-1.5 w-1.5 rounded-full bg-emerald-500"></div>
+                                        </div>
+                                        <p class="text-xs leading-5 text-zinc-900 dark:text-zinc-400">Online</p>
+                                    </div>
+                                </div>
+                            </li>
+                        </ul>
+                    </div>
+                </div>
+            </div>
+        </article>
+    </main>
+</template>
